+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/sweep/BackgroundSweeperImpl.java
@@ -46,7 +46,6 @@ import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.atlasdb.keyvalue.api.InsufficientConsistencyException;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
-import com.palantir.atlasdb.keyvalue.api.RangeRequest;
 import com.palantir.atlasdb.keyvalue.api.RowResult;
 import com.palantir.atlasdb.keyvalue.api.SweepResults;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
@@ -354,8 +353,9 @@ public class BackgroundSweeperImpl implements BackgroundSweeper {
                             .build());
         }
 
-        // use delete instead of truncate so that sweep does not interfere with dbkvs online backups
-        kvs.deleteRange(tableFactory.getSweepProgressTable(null).getTableRef(), RangeRequest.all());
+        // Truncate instead of delete because the progress table contains only
+        // a single row that has accumulated many overwrites.
+        kvs.truncateTable(tableFactory.getSweepProgressTable(null).getTableRef());
     }
 
     private void saveIntermediateSweepResults(final SweepProgressRowResult progress,
@@ -422,8 +422,7 @@ public class BackgroundSweeperImpl implements BackgroundSweeper {
             if (result == null || tables.contains(result.getFullTableName())) {
                 return false;
             }
-            // use delete instead of truncate so that sweep does not interfere with dbkvs online backups
-            kvs.deleteRange(tableFactory.getSweepProgressTable(null).getTableRef(), RangeRequest.all());
+            kvs.truncateTable(tableFactory.getSweepProgressTable(null).getTableRef());
             return true;
         } catch (RuntimeException e) {
             log.error("Failed to check whether the table being swept was dropped."